pagina_web/requirements.txt
Flask
psycopg2-binary
Werkzeug
python-dotenv
PyPDF2
PyMuPDF
pdf2image
pytesseract

#Flask==3.0.2
#psycopg2-binary==2.9.9
#Werkzeug==3.0.1
#python-dotenv==1.0.1
#PyPDF2==3.0.1
#PyMuPDF==1.23.26
#pdf2image==1.17.0
#pytesseract==0.3.10